Cargo bounty fix requests
