+++ Content.Shared/Cargo/Components/CargoBountyConsoleComponent.cs
@@ -95,28 +95,3 @@ public sealed class BountySkipMessage : BoundUserInterfaceMessage
     }
 }
 
-// Begin DeltaV bounty claiming
-[Serializable, NetSerializable]
-public sealed class BountyClaimedMessage : BoundUserInterfaceMessage
-{
-    public string BountyId;
-
-    public BountyClaimedMessage(string bountyId)
-    {
-        BountyId = bountyId;
-    }
-}
-
-[Serializable, NetSerializable]
-public sealed class BountySetStatusMessage : BoundUserInterfaceMessage
-{
-    public string BountyId;
-    public int Status;
-
-    public BountySetStatusMessage(string bountyId, int status)
-    {
-        BountyId = bountyId;
-        Status = status;
-    }
-}
-// End DeltaV bounty claiming
